requirements.txt
Flask
gunicorn
python-dotenv
google-cloud-texttospeech
waitress
